test: move src inline tests to tests/ and fix CI no-tty failures
- Add atty guards to terminal_component_get_creates_terminal_backend
  and test_factory_create_registers_di_screens (Linux CI has no TTY).
- Remove all `#[cfg(test)] mod tests` blocks from src/, relocate the
  coverage to tests/ via `#[cfg(feature = "test-mocks")] pub fn
  ..._for_test(...)` helpers on the corresponding types.
- Add AGENTS.md documenting the "no tests in src/" rule and the
  no-TTY CI constraint.
