+++ src/domain/services/source_code_parser/parsers/cpp.rs
@@ -303,92 +303,3 @@ impl CppExtractor {
         None
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    fn parse(source_code: &str) -> tree_sitter::Tree {
-        let mut parser = CppExtractor::create_parser().unwrap();
-        parser.parse(source_code, None).unwrap()
-    }
-
-    fn find_node<'tree>(
-        node: Node<'tree>,
-        source_code: &str,
-        kind: &str,
-        text: &str,
-    ) -> Option<Node<'tree>> {
-        let matches_kind = node.kind() == kind;
-        let matches_text = node
-            .utf8_text(source_code.as_bytes())
-            .map(|node_text| node_text.contains(text))
-            .unwrap_or(false);
-
-        matches_kind
-            .then_some(node)
-            .filter(|_| matches_text)
-            .or_else(|| {
-                (0..node.child_count())
-                    .filter_map(|index| node.child(index))
-                    .find_map(|child| find_node(child, source_code, kind, text))
-            })
-    }
-
-    #[test]
-    fn extract_name_returns_none_for_cpp_namespace_identifier_nodes() {
-        let source_code = "namespace tools { int value = 1; }";
-        let tree = parse(source_code);
-        let namespace_node = find_node(
-            tree.root_node(),
-            source_code,
-            "namespace_definition",
-            "namespace tools",
-        )
-        .unwrap();
-
-        let name = CppExtractor.extract_name(namespace_node, source_code, "namespace.definition");
-
-        assert_eq!(name, None);
-    }
-
-    #[test]
-    fn extract_name_reads_constructor_and_handles_destructor_fallback() {
-        let source_code = "class Resource { public: Resource() {} ~Resource() {} };";
-        let tree = parse(source_code);
-        let constructor_node = find_node(
-            tree.root_node(),
-            source_code,
-            "function_definition",
-            "Resource() {}",
-        )
-        .unwrap();
-        let destructor_node = find_node(
-            tree.root_node(),
-            source_code,
-            "function_definition",
-            "~Resource() {}",
-        )
-        .unwrap();
-
-        let constructor_name =
-            CppExtractor.extract_name(constructor_node, source_code, "constructor.definition");
-        let destructor_name =
-            CppExtractor.extract_name(destructor_node, source_code, "destructor.definition");
-
-        assert_eq!(constructor_name, Some("Resource".to_string()));
-        assert_eq!(destructor_name, None);
-    }
-
-    #[test]
-    fn extract_name_reads_plain_variable_declarations() {
-        let source_code = "int plain;";
-        let tree = parse(source_code);
-        let declaration_node =
-            find_node(tree.root_node(), source_code, "declaration", "plain").unwrap();
-
-        let name = CppExtractor.extract_name(declaration_node, source_code, "variable.definition");
-
-        assert_eq!(name, Some("plain".to_string()));
-    }
-}
