+++ src/presentation/tui/screens/total_summary_share_screen.rs
@@ -139,6 +139,26 @@ impl TotalSummaryShareScreen {
         browser::open_url(url)
             .map_err(|e| GitTypeError::TerminalError(format!("Failed to open browser: {}", e)))
     }
+
+    #[cfg(feature = "test-mocks")]
+    pub fn generate_share_url_for_test(&self, text: &str, platform: &SharingPlatform) -> String {
+        self.generate_share_url(text, platform)
+    }
+
+    #[cfg(feature = "test-mocks")]
+    pub fn set_total_result_for_test(&self, total_result: TotalResult) {
+        *self.total_result.write().unwrap() = total_result;
+    }
+
+    #[cfg(feature = "test-mocks")]
+    pub fn set_fallback_url_for_test(&self, value: Option<(String, SharingPlatform)>) {
+        *self.fallback_url.write().unwrap() = value;
+    }
+
+    #[cfg(feature = "test-mocks")]
+    pub fn fallback_url_for_test(&self) -> Option<(String, SharingPlatform)> {
+        self.fallback_url.read().unwrap().clone()
+    }
 }
 
 pub struct TotalSummaryShareScreenProvider;
@@ -241,69 +261,3 @@ impl Screen for TotalSummaryShareScreen {
 }
 
 impl TotalSummaryShareScreenInterface for TotalSummaryShareScreen {}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::domain::events::EventBus;
-    use crate::domain::models::color_mode::ColorMode;
-    use crate::domain::models::theme::Theme;
-    use crate::domain::services::scoring::TotalTracker;
-    use crate::domain::services::theme_service::ThemeService;
-
-    fn make_screen() -> TotalSummaryShareScreen {
-        TotalSummaryShareScreen::new(
-            Arc::new(EventBus::new()),
-            Arc::new(ThemeService::new_for_test(
-                Theme::default(),
-                ColorMode::Dark,
-            )) as Arc<dyn ThemeServiceInterface>,
-            Arc::new(TotalTracker::new_for_test()) as Arc<dyn TotalTrackerInterface>,
-        )
-    }
-
-    #[test]
-    fn generate_share_url_encodes_each_platform() {
-        let screen = make_screen();
-        let mut total_result = TotalResult::new();
-        total_result.total_keystrokes = 1234;
-        total_result.total_score = 9876.0;
-        total_result.overall_cpm = 321.0;
-        *screen.total_result.write().unwrap() = total_result;
-
-        let text = "Total score: 9876 & CPM: 321";
-
-        let x_url = screen.generate_share_url(text, &SharingPlatform::X);
-        assert!(x_url.starts_with("https://x.com/intent/tweet?text="));
-        assert!(x_url.contains("Total%20score%3A%209876%20%26%20CPM%3A%20321"));
-
-        let reddit_url = screen.generate_share_url(text, &SharingPlatform::Reddit);
-        assert!(reddit_url.starts_with("https://www.reddit.com/submit?"));
-        assert!(reddit_url.contains("title=Just%20demolished%201234%20keystrokes"));
-        assert!(reddit_url.contains("selftext=true"));
-        assert!(reddit_url.contains("text=Total%20score%3A%209876%20%26%20CPM%3A%20321"));
-
-        let linked_in_url = screen.generate_share_url(text, &SharingPlatform::LinkedIn);
-        assert!(linked_in_url.starts_with("https://www.linkedin.com/feed/"));
-        assert!(linked_in_url.contains("shareActive=true"));
-        assert!(linked_in_url.contains("mini=true"));
-
-        let facebook_url = screen.generate_share_url(text, &SharingPlatform::Facebook);
-        assert!(facebook_url.starts_with("https://www.facebook.com/sharer/sharer.php?"));
-        assert!(facebook_url.contains("u=https%3A%2F%2Fgithub.com%2Funhappychoice%2Fgittype"));
-        assert!(facebook_url.contains("quote=Total%20score%3A%209876%20%26%20CPM%3A%20321"));
-    }
-
-    #[test]
-    fn escape_clears_fallback_url_before_navigating_back() {
-        let screen = make_screen();
-        *screen.fallback_url.write().unwrap() =
-            Some(("https://example.test/share".to_string(), SharingPlatform::X));
-
-        screen
-            .handle_key_event(event::KeyEvent::new(KeyCode::Esc, KeyModifiers::empty()))
-            .unwrap();
-
-        assert!(screen.fallback_url.read().unwrap().is_none());
-    }
-}
