+++ tests/unit/presentation/mod.rs
@@ -4,5 +4,6 @@ pub mod cli_screen_runner_tests;
 pub mod cli_trending_tests;
 pub mod game;
 pub mod sharing_tests;
+pub mod signal_handler_tests;
 pub mod tui;
 pub mod ui;
