+++ tests/unit/presentation/tui/mod.rs
@@ -4,6 +4,7 @@ pub mod screen_manager_tests;
 pub mod screen_trait_tests;
 pub mod screen_transition_manager_tests;
 pub mod total_summary_ascii_score_view_tests;
+pub mod total_summary_share_screen_tests;
 pub mod total_summary_share_sharing_view_tests;
 pub mod typing_animation_view_tests;
 pub mod typing_countdown_view_tests;
